pyproject.toml
[tool.poetry]
name = "teleddns-server"
version = "0.3.0"
description = "DDNS API server"
authors = ["Tomas Hlavacek <tmshlvck@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.2.0"
djangorestframework = "^3.16.1"
django-environ = "^0.11.0"
requests = "^2.31.0"
dnspython = "^2.4.0"
pyotp = "^2.9.0"
django-cors-headers = "^4.3.0"
pwdlib = {version = "^0.2.0", extras = ["argon2"]}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.8.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.27.0"
requests = "^2.31.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

